\name{trip_mpg}
\alias{trip_mpg}
\title{Trip mpg}
\usage{
trip_mpg()
}
\description{
Calculates a trip's cost based on 3 factors:
miles - total number of miles traveled for a particular trip
mpg - miles per galon value in a car traveled
gasprice - current gas price for that particular car
}

\examples{
trip_mpg(25,28.2,4.0)
}
